DEV-3937, DEV-3932, DEV-3956, fixes
git-svn-id: https://svn.terracotta.org/repo/tc/dso/trunk@14774 7fc7bbf3-cf45-46d4-be06-341739edd864

+++ code/base/dso-l2/src/com/tc/util/tickertoken/TickerTokenManagerImpl.java
@@ -18,7 +18,7 @@ import java.util.concurrent.atomic.AtomicBoolean;
 public abstract class TickerTokenManagerImpl implements TickerTokenManager {
 
   private final int                                                  id;
-  private final int                                                  timerPeriod;
+  private final long                                                 timerPeriod;
 
   private final Map<Class, TickerTokenFactory>                       factoryMap         = Collections
                                                                                             .synchronizedMap(new HashMap<Class, TickerTokenFactory>());
@@ -38,7 +38,7 @@ public abstract class TickerTokenManagerImpl implements TickerTokenManager {
                                                                                                     "Ticker Token Timer",
                                                                                                     true);
 
-  public TickerTokenManagerImpl(int id, int totalTickers, int timerPeriod) {
+  public TickerTokenManagerImpl(int id, int totalTickers, long timerPeriod) {
     this.id = id;
     this.totalTickers = totalTickers;
     this.timerPeriod = timerPeriod;
